requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
gunicorn==20.1.0
reportlab==4.0.4
